# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Moxxy Agentic Framework API
  version: 0.1.0
  description: Local-first agentic framework REST API
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "mox_*"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
...
    AgentUpdate:
      type: object
      properties:
        provider_id:
          type: string
        model_id:
          type: string
        temperature:
          type: number
          minimum: 0.0
          maximum: 2.0
        persona:
          type: string
        template:
          type: string
          nullable: true
          description: Template slug (null to clear)
    AgentResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        provider_id:
          type: string
        model_id:
          type: string
        status:
          type: string
          enum: [idle, running, stopping]
        persona:
          type: string
          nullable: true
        template:
          type: string
          nullable: true
          description: Assigned template slug
        created_at:
          type: string
...
paths:
...
  /v1/agents/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Agents]
      summary: Get agent
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags: [Agents]
      summary: Update agent config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentUpdate'
      responses:
        '200':
          description: Agent updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                    nullable: true
                  provider_id:
                    type: string
                  model_id:
                    type: string
                  temperature:
                    type: number
                  status:
                    type: string
                  updated_at:
                    type: string
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Agents]
      summary: Delete agent
      responses:
        '204':
          description: Agent deleted
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...